Divide preallocated module from exposer to resolve circular imports
