--- apiserver/paasng/paasng/engine/configurations/__init__.py
@@ -1,8 +1,0 @@
-"""Store configurations related with deployment"""
-from .config_var import get_env_variables
-from .image import ImageCredentialManager
-
-__all__ = [
-    'get_env_variables',
-    'ImageCredentialManager',
-]
