+++ apiserver/paasng/paasng/engine/deploy/steps/pre_release.py
@@ -21,7 +21,7 @@ from typing import Dict
 from blue_krill.async_utils.poll_task import CallbackHandler, CallbackResult, PollingResult, PollingStatus, TaskPoller
 from django.utils.translation import gettext as _
 
-from paasng.engine.configurations import get_env_variables
+from paasng.engine.configurations.config_var import get_env_variables
 from paasng.engine.constants import JobStatus
 from paasng.engine.deploy.engine_svc import EngineDeployClient
 from paasng.engine.deploy.infra.models_utils import update_engine_app_config
