+++ apiserver/paasng/paasng/publish/entrance/utils.py
@@ -17,15 +17,18 @@ We undertake not to change the open source license (MIT license) applicable
 to the current version of the project delivered to anyone in the future.
 """
 from dataclasses import dataclass
+from typing import NamedTuple, Optional
 from urllib.parse import urlparse
 
 from paas_wl.cluster.models import PortMap
+from paasng.platform.applications.models import ModuleEnvironment
+from paasng.platform.region.models import get_region
 
 default_port_map = PortMap()
 
 
 def to_dns_safe(s: str) -> str:
-    """Tranform some string to dns safe"""
+    """Transform some string to dns safe"""
     return s.replace('_', '--').lower()
 
 
@@ -51,3 +54,24 @@ class URL:
         port = parsed.port or default_port_map.get_port_num(protocol)
         assert parsed.hostname
         return URL(protocol=protocol, hostname=parsed.hostname, port=port, path=parsed.path, query=parsed.query)
+
+
+class EnvExposedURL(NamedTuple):
+    url: URL
+    provider_type: str
+
+    @property
+    def address(self) -> str:
+        return self.url.as_address()
+
+
+def get_legacy_url(env: ModuleEnvironment) -> Optional[str]:
+    """Deprecated: Get legacy URL address which is a hard-coded value generated
+    y region configuration.
+
+    :return: None if not configured.
+    """
+    app = env.application
+    if tmpl := get_region(app.region).basic_info.link_engine_app:
+        return tmpl.format(code=app.code, region=app.region, name=env.engine_app.name)
+    return None
