+++ apiserver/paasng/tests/engine/configurations/test_config_var.py
@@ -22,11 +22,14 @@ from unittest import mock
 
 import pytest
 from blue_krill.contextlib import nullcontext as does_not_raise
+from django.conf import settings
 
-from paasng.engine.configurations.config_var import get_env_variables
+from paasng.engine.configurations.config_var import get_builtin_env_variables, get_env_variables
+from paasng.engine.constants import AppRunTimeBuiltinEnv
 from paasng.engine.models.config_var import ConfigVar
 from paasng.extensions.declarative.exceptions import DescriptionValidationError
 from paasng.extensions.declarative.handlers import AppDescriptionHandler
+from tests.utils.helpers import override_region_configs
 from tests.utils.mocks.engine import mock_cluster_service
 
 pytestmark = pytest.mark.django_db
@@ -130,3 +133,39 @@ class TestGetEnvVariables:
         AppDescriptionHandler.from_file(fp).handle_deployment(bk_deployment)
         env_vars = get_env_variables(bk_stag_env, deployment=bk_deployment)
         assert 'BKPAAS_SERVICE_ADDRESSES_BKSAAS' in env_vars
+
+
+class TestBuiltInEnvVars:
+    @pytest.mark.parametrize("provide_env_vars_platform, contain_bk_envs", [(True, True), (False, True)])
+    def test_bk_platform_envs(self, bk_app, provide_env_vars_platform, contain_bk_envs):
+        def update_region_hook(config):
+            config['provide_env_vars_platform'] = provide_env_vars_platform
+
+        with override_region_configs(bk_app.region, update_region_hook):
+            bk_module = bk_app.get_default_module()
+            bk_stag_env = bk_module.envs.get(environment='stag')
+            config_vars = get_builtin_env_variables(bk_stag_env.engine_app, settings.CONFIGVAR_SYSTEM_PREFIX)
+
+            # 这些环境变量在所有版本都有
+            assert ('BK_COMPONENT_API_URL' in config_vars) == contain_bk_envs
+            assert ('BK_PAAS2_URL' in config_vars) == contain_bk_envs
+            assert ('BK_API_URL_TMPL' in config_vars) == contain_bk_envs
+
+            # BK_LOGIN_URL 只在特殊开启的版本才写入
+            assert ('BK_LOGIN_URL' in config_vars) == provide_env_vars_platform
+            # 应用是需要写入蓝鲸体系其他系统访问地址的环境变量
+            if provide_env_vars_platform:
+                assert set(settings.BK_PLATFORM_URLS.keys()).issubset(set(config_vars.keys())) == contain_bk_envs
+
+    def test_builtin_env_keys(self, bk_app):
+        bk_module = bk_app.get_default_module()
+        bk_stag_env = bk_module.envs.get(environment='stag')
+        config_vars = get_builtin_env_variables(bk_stag_env.engine_app, settings.CONFIGVAR_SYSTEM_PREFIX)
+
+        assert {'BKPAAS_LOGIN_URL', 'BKPAAS_APP_CODE', 'BKPAAS_APP_ID', 'BKPAAS_APP_SECRET'}.issubset(
+            config_vars.keys()
+        )
+
+        # 运行时相关的环境变量
+        runtime_env_keys = [f'{settings.CONFIGVAR_SYSTEM_PREFIX}{key}' for key in AppRunTimeBuiltinEnv.get_values()]
+        assert set(runtime_env_keys).issubset(config_vars.keys())
